+++ ubiq_security/structured/decrypt.py
@@ -6,15 +6,15 @@ from ..credentials import credentials
 
 
 from .common import fmtInput, strConvertRadix, decKeyNumber, fmtOutput
-from .common import fetchFFS, fetchKey
-from ubiq_security_fpe import ff1
+from .common import fetchDataset, fetchKey
+from .lib import ff1
 
 class Decryption:
     def __del__(self):
         # todo: overwrite 'unwrapped_data_key'
         return
 
-    def __init__(self, creds, ffs):
+    def __init__(self, creds, dataset_name):
         if not creds.set():
             raise RuntimeError("credentials not set")
         
@@ -31,35 +31,35 @@ class Decryption:
         self._sapi = creds.secret_signing_key
         self._srsa = creds.secret_crypto_access_key
 
-        self._ffs = fetchFFS(self._host, self._papi, self._sapi, ffs)
+        self._dataset = fetchDataset(self._host, self._papi, self._sapi, dataset_name)
 
     def Cipher(self, ct, twk = None):
-        pth = self._ffs['passthrough']
-        ics = self._ffs['input_character_set']
-        ocs = self._ffs['output_character_set']
+        pth = self._dataset['passthrough']
+        ics = self._dataset['input_character_set']
+        ocs = self._dataset['output_character_set']
 
         fmt, ct = fmtInput(ct, pth, ocs, ics)
-        ct, n = decKeyNumber(ct, ocs, self._ffs['msb_encoding_bits'])
+        ct, n = decKeyNumber(ct, ocs, self._dataset['msb_encoding_bits'])
         if not hasattr(self, '_key') or self._key['key_number'] != n:
             self._key = fetchKey(self._host,
                                  self._papi, self._sapi, self._srsa,
-                                 self._ffs['name'], n)
-            if self._ffs['encryption_algorithm'] == 'FF1':
+                                 self._dataset['name'], n)
+            if self._dataset['encryption_algorithm'] == 'FF1':
                 self._ctx = ff1.Context(
                     self._key['unwrapped_data_key'],
-                    base64.b64decode(self._ffs['tweak']),
-                    self._ffs['tweak_min_len'], self._ffs['tweak_max_len'],
+                    base64.b64decode(self._dataset['tweak']),
+                    self._dataset['tweak_min_len'], self._dataset['tweak_max_len'],
                     len(ics), ics)
             else:
                 raise RuntimeError('unsupported algorithm: ' +
-                                   self._ffs['encryption_algorithm'])
+                                   self._dataset['encryption_algorithm'])
         ct = strConvertRadix(ct, ocs, ics)
 
         pt = self._ctx.Decrypt(ct, twk)
 
-        self._creds.add_event(dataset_name=self._ffs['name'], dataset_group_name="", billing_action="decrypt",
+        self._creds.add_event(dataset_name=self._dataset['name'], dataset_group_name="", billing_action="decrypt",
             dataset_type="structured", key_number=n, count=1)
         return fmtOutput(fmt, pt, pth)
 
-def Decrypt(creds, ffs, ct, twk = None):
-    return Decryption(creds, ffs).Cipher(ct, twk)
+def Decrypt(creds, dataset_name, ct, twk = None):
+    return Decryption(creds, dataset_name).Cipher(ct, twk)